+++ tools/resource/mesh_resource_obj.vala
@@ -92,6 +92,7 @@ namespace OBJImport
 		, bool create_textures_folder
 		, ufbx.MaterialMap map
 		, TextureUsage usage
+		, bool preserve_alpha
 		, Gee.HashMap<unowned ufbx.Texture, string> imported_textures
 		)
 	{
@@ -100,7 +101,7 @@ namespace OBJImport
 			return 0;
 
 		unowned ufbx.Texture? texture = map.texture;
-		if (imported_textures.has_key(texture)) {
+		if (imported_textures.has_key(texture) && !preserve_alpha) {
 			resource_name = imported_textures[texture];
 			return 0;
 		}
@@ -163,6 +164,8 @@ namespace OBJImport
 			texture_resource = TextureResource.normal_map(db, texture_id, source_image);
 		else if ((usage & TextureUsage.DATA) != 0)
 			texture_resource = TextureResource.data_map(db, texture_id, source_image);
+		else if (preserve_alpha)
+			texture_resource = TextureResource(db, texture_id, source_image, TextureFormat.BC3, true, false);
 		else
 			texture_resource = TextureResource.color_map(db, texture_id, source_image);
 		if (texture_resource.save(project, texture_resource_name) != 0)
@@ -613,6 +616,7 @@ public class OBJImporter
 					string? roughness_map = null;
 					string? ao_map = null;
 					string? emission_map = null;
+					bool masking = false;
 
 					for (int mm = 0; mm < ufbx.MaterialPbrMap.MAP_COUNT; ++mm) {
 						unowned ufbx.MaterialMap map = material.pbr.maps[mm];
@@ -621,6 +625,20 @@ public class OBJImporter
 						case ufbx.MaterialPbrMap.BASE_COLOR: {
 								if (map.has_value)
 									albedo = OBJImport.vector3(map.value_vec3);
+
+								unowned ufbx.MaterialMap opacity = material.pbr.opacity;
+								masking = map.texture_enabled
+									&& opacity.texture_enabled
+									&& map.texture != null
+									&& opacity.texture != null
+									;
+								if (masking) {
+									string color_texture_filename = OBJImport.texture_filename(map.texture);
+									string opacity_texture_filename = OBJImport.texture_filename(opacity.texture);
+									masking = map.texture == opacity.texture
+										|| (color_texture_filename.length > 0 && color_texture_filename == opacity_texture_filename)
+										;
+								}
 
 								if (options.import_textures.value
 									&& OBJImport.get_or_import_texture_resource_name(out albedo_map
@@ -632,6 +650,7 @@ public class OBJImporter
 									, options.create_textures_folder.value
 									, map
 									, TextureUsage.COLOR
+									, masking
 									, imported_textures
 									) != 0)
 									return ImportResult.ERROR;
@@ -649,6 +668,7 @@ public class OBJImporter
 								, options.create_textures_folder.value
 								, map
 								, TextureUsage.NORMAL
+								, false
 								, imported_textures
 								) != 0)
 								return ImportResult.ERROR;
@@ -668,6 +688,7 @@ public class OBJImporter
 								, options.create_textures_folder.value
 								, map
 								, TextureUsage.DATA
+								, false
 								, imported_textures
 								) != 0)
 								return ImportResult.ERROR;
@@ -687,6 +708,7 @@ public class OBJImporter
 								, options.create_textures_folder.value
 								, map
 								, TextureUsage.DATA
+								, false
 								, imported_textures
 								) != 0)
 								return ImportResult.ERROR;
@@ -703,6 +725,7 @@ public class OBJImporter
 								, options.create_textures_folder.value
 								, map
 								, TextureUsage.DATA
+								, false
 								, imported_textures
 								) != 0)
 								return ImportResult.ERROR;
@@ -722,6 +745,7 @@ public class OBJImporter
 								, options.create_textures_folder.value
 								, map
 								, TextureUsage.COLOR
+								, false
 								, imported_textures
 								) != 0)
 								return ImportResult.ERROR;
@@ -736,6 +760,8 @@ public class OBJImporter
 							break;
 						}
 					}
+
+					masking = masking && albedo_map != null;
 
 					// Create .material resource.
 					MaterialResource material_resource = MaterialResource.mesh(db
@@ -752,6 +778,7 @@ public class OBJImporter
 						, emission_color
 						, emission_intensity
 						, shader
+						, masking
 						);
 					if (material_resource.save(project, material_resource_name) != 0)
 						return ImportResult.ERROR;